refactor(ios): split shared chat state owners
