+++ apps/shared/OpenClawKit/Sources/OpenClawChatUI/ChatViewModel+Thinking.swift
@@ -5,6 +5,14 @@ import Foundation
 // extension owns collapsing them into the canonical option list.
 
 extension OpenClawChatViewModel {
+    static let baseThinkingLevelOptions: [OpenClawChatThinkingLevelOption] = [
+        OpenClawChatThinkingLevelOption(id: "off", label: "off"),
+        OpenClawChatThinkingLevelOption(id: "minimal", label: "minimal"),
+        OpenClawChatThinkingLevelOption(id: "low", label: "low"),
+        OpenClawChatThinkingLevelOption(id: "medium", label: "medium"),
+        OpenClawChatThinkingLevelOption(id: "high", label: "high"),
+    ]
+
     func applyAdvertisedThinkingLevel(_ level: String) {
         guard level != thinkingLevel else { return }
         thinkingLevel = level
