+++ apps/shared/OpenClawKit/Sources/OpenClawChatUI/ChatViewModel.swift
@@ -755,13 +755,6 @@ public final class OpenClawChatViewModel {
         return "Default: \(modelLabel(for: defaultModelID))"
     }
 
-    static let baseThinkingLevelOptions: [OpenClawChatThinkingLevelOption] = [
-        OpenClawChatThinkingLevelOption(id: "off", label: "off"),
-        OpenClawChatThinkingLevelOption(id: "minimal", label: "minimal"),
-        OpenClawChatThinkingLevelOption(id: "low", label: "low"),
-        OpenClawChatThinkingLevelOption(id: "medium", label: "medium"),
-        OpenClawChatThinkingLevelOption(id: "high", label: "high"),
-    ]
 }
 
 extension OpenClawChatViewModel {
